+++ jmix-localized-string-datatype/src/test/java/com/glebfox/jmix/locstr/entity/LocalizedStringTestEntity.java
@@ -18,11 +18,11 @@ package com.glebfox.jmix.locstr.entity;
 
 import com.glebfox.jmix.locstr.datatype.LocalizedString;
 import com.glebfox.jmix.locstr.validation.constraints.LocalizedStringNotBlank;
+import com.glebfox.jmix.locstr.validation.constraints.LocalizedStringSize;
 import io.jmix.core.entity.annotation.JmixGeneratedValue;
 import io.jmix.core.metamodel.annotation.InstanceName;
 import io.jmix.core.metamodel.annotation.JmixEntity;
 import io.jmix.core.metamodel.annotation.JmixProperty;
-import io.jmix.core.validation.group.UiComponentChecks;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
@@ -30,6 +30,8 @@ import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
 import jakarta.persistence.Version;
 
+import java.util.Locale;
+import java.util.Map;
 import java.util.UUID;
 
 @JmixEntity
@@ -50,10 +52,15 @@ public class LocalizedStringTestEntity {
     @Column(name = "NAME", nullable = false)
     private LocalizedString name;
 
-    @LocalizedStringNotBlank(groups = UiComponentChecks.class)
+    @LocalizedStringNotBlank
     @JmixProperty
     @Transient
-    private LocalizedString validationName;
+    private LocalizedString validationName = new LocalizedString(Map.of(Locale.ENGLISH, "value"));
+
+    @LocalizedStringSize(max = 100)
+    @JmixProperty
+    @Transient
+    private LocalizedString validationOptional;
 
     public UUID getId() {
         return id;
@@ -87,5 +94,15 @@ public class LocalizedStringTestEntity {
     @SuppressWarnings("unused")
     public void setValidationName(LocalizedString validationName) {
         this.validationName = validationName;
+    }
+
+    @SuppressWarnings("unused")
+    public LocalizedString getValidationOptional() {
+        return validationOptional;
+    }
+
+    @SuppressWarnings("unused")
+    public void setValidationOptional(LocalizedString validationOptional) {
+        this.validationOptional = validationOptional;
     }
 }